# zenzefi_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ZenZefi REST API provides operations for resources certificate, configuration, log, user, system",
        "version": "1.0",
        "title": "ZenZefi REST API"
    },
    "host": "localhost:61000",
    "basePath": "/",
...
    "paths": {
        "/certificates": {
            "get": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi Certificate Management"
                ],
                "summary": "Returns an array of certificates belonging to the current logged in user.",
                "operationId": "allUsingGET_1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ZenZefiCertificate"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi Certificate Management"
                ],
                "summary": "Deletes all certificates from current user-specific certificate store, or by authority key identifier and serial number. For details about the parameters, please view the request model.",
                "operationId": "deleteCertificates_1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "all",
                        "in": "query",
                        "description": "If set to true, deletes all certificates.",
                        "required": false,
                        "type": "boolean",
                        "default": false
                    },
                    {
                        "in": "body",
                        "name": "pairs",
                        "description": "Base64 encoded authority key identifier / serial number pairs.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DeleteCertificateModel"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DeleteCertificatesResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Empty input list. | Invalid authority key identifier. The length is 20 bytes. | The input is not Base64 encoded. | Invalid serial number. The max length is 16 bytes. | The certificate with id does not exist in current user-specific certificate store."
                    },
                    "404": {
                        "description": "Delete certificates called but no match was found."
                    }
                }
            }
        },
...
        "/certificates/checkSystemIntegrityLog": {
            "get": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi Certificate Management"
                ],
                "summary": "Returns the system integrity check XML report. Endpoint used only in the ZenZefi UI.",
                "operationId": "getIntegrityCheckLogFile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/File"
                        }
                    },
                    "400": {
                        "description": "System integrity check xml report not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/certificates/{ids}": {
            "delete": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi Certificate Management"
                ],
                "summary": "Deletes one or more certificates from current user-specific certificate store. The deletion is made by certificate ids, and each certificate id must be a valid UUID. The ids are comma separated.",
                "operationId": "deleteCertificates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "The ids of the certificates, comma separated. Must be valid UUIDs.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DeleteCertificatesResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid format for the given ID. | Delete certificates called but no match was found. | The certificate with id does not exist in current user-specific certificate store."
                    }
                }
            }
        },
        "/certificates/{id}": {
            "get": {
                "tags": [
                    "ZenZefi Certificate Management"
                ],
                "summary": "Returns a json representation of required certificate by id. The id must be a valid UUID.",
                "operationId": "getCertificateUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*",
                    "JSON"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the certificate. Must be valid UUID.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Certificate"
                        }
                    },
                    "400": {
                        "description": "Certificate does not exist | Invalid format for the given ID."
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/v1/certificates": {
            "get": {
                "tags": [
                    "ZenZefi Public API V1",
                    "ZenZefi Certificate Management"
                ],
                "summary": "Returns an array of certificates belonging to the current logged in user.",
                "operationId": "allV1UsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ZenZefiCertificate"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ZenZefi Public API V1",
                    "ZenZefi Certificate Management"
                ],
                "summary": "Deletes all certificates from current user-specific certificate store, or by authority key identifier and serial number. For details about the parameters, please view the request model.",
                "operationId": "deleteCertificatesV1_1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "all",
                        "in": "query",
                        "description": "If set to true, deletes all certificates.",
                        "required": false,
                        "type": "boolean",
                        "default": false
                    },
                    {
                        "in": "body",
                        "name": "pairs",
                        "description": "Base64 encoded authority key identifier / serial number pairs.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DeleteCertificateModel"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DeleteCertificatesResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Empty input list. | Invalid authority key identifier. The length is 20 bytes. | The input is not Base64 encoded. | Invalid serial number. The max length is 16 bytes. | The certificate with id does not exist in current user-specific certificate store."
                    },
                    "404": {
                        "description": "Delete certificates called but no match was found."
                    }
                }
            }
        },
...
        "/v1/certificates/checkSystemIntegrityLog": {
            "get": {
                "tags": [
                    "ZenZefi Public API V1",
                    "ZenZefi Certificate Management"
                ],
                "summary": "Returns the system integrity check XML report. Endpoint used only in the ZenZefi UI.",
                "operationId": "getIntegrityCheckLogFileV1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/File"
                        }
                    },
                    "400": {
                        "description": "System integrity check xml report not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/v1/certificates/{ids}": {
            "delete": {
                "tags": [
                    "ZenZefi Public API V1",
                    "ZenZefi Certificate Management"
                ],
                "summary": "Deletes one or more certificates from current user-specific certificate store. The deletion is made by certificate ids, and each certificate id must be a valid UUID. The ids are comma separated.",
                "operationId": "deleteCertificatesV1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "The ids of the certificates, comma separated. Must be valid UUIDs.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DeleteCertificatesResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid format for the given ID. | The certificate with id does not exist in current user-specific certificate store."
                    },
                    "404": {
                        "description": "Delete certificates called but no match was found."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "DeleteCertificateModel": {
            "type": "object",
            "required": [
                "authorityKeyIdentifier",
                "serialNo"
            ],
            "properties": {
                "authorityKeyIdentifier": {
                    "type": "string",
                    "description": "The authority key identifier. It is sent as Base64 encoded bytes, and the length must be 20 bytes."
                },
                "serialNo": {
                    "type": "string",
                    "description": "The serial number. It is sent as Base64 encoded bytes. The maximum length is 16 bytes."
                }
            }
        },
        "DeleteCertificatesResult": {
            "type": "object",
            "properties": {
                "certificateType": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "File": {
            "type": "object",
            "properties": {
                "absolute": {
                    "type": "boolean"
                },
                "absoluteFile": {
                    "$ref": "#/definitions/File"
                },
                "absolutePath": {
                    "type": "string"
                },
                "canonicalFile": {
                    "$ref": "#/definitions/File"
                },
                "canonicalPath": {
                    "type": "string"
                },
                "directory": {
                    "type": "boolean"
                },
                "file": {
                    "type": "boolean"
                },
                "freeSpace": {
                    "type": "integer",
                    "format": "int64"
                },
                "hidden": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "type": "string"
                },
                "parentFile": {
                    "$ref": "#/definitions/File"
                },
                "path": {
                    "type": "string"
                },
                "totalSpace": {
                    "type": "integer",
                    "format": "int64"
                },
                "usableSpace": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
    }
}